Fixed truncation warnings; Using floats instead of doubles in all instances

+++ 3501Project/src/AssetManager.cpp
@@ -29,7 +29,7 @@ namespace AsteroidGame{
 		//Max Note: Magic number above. Do we just want shield to have a default radius value?
 	}
 
-	Ogre::SceneNode* AssetManager::createAsteroid(double aRadius) {
+	Ogre::SceneNode* AssetManager::createAsteroid(float aRadius) {
 		return iResourceFactory->createAsteroidModel(iApplication->getSceneManager(), aRadius);
 	}
